minor, moving a util from core xformers to LRA
